Introduce ChannelList to hold channel collection
