+++ lib/worker-connection.js
@@ -7,18 +7,18 @@ var WorkerUtil = require('./worker-util');
 
 /**
  * @param {SockJSConnection} connection
- * @param {Worker} worker
+ * @param {ChannelList} channelList
  * @constructor
  */
-function WorkerConnection(connection, worker) {
+function WorkerConnection(connection, channelList) {
   /** @type {String} */
   this.id = connection.id;
 
   /** @type {SockJSConnection} */
   this._connection = connection;
 
-  /** @type {Worker} */
-  this._worker = worker;
+  /** @type {ChannelList} */
+  this._channelList = channelList;
 
   /** @type {String[]} */
   this._сhannelIds = [];
@@ -67,10 +67,10 @@ WorkerConnection.prototype.subscribe = function(channelId, data, msgStartTime) {
   }
   msgStartTime = msgStartTime || new Date().getTime();
   this.addChannelId(channelId);
-  if (!this._worker.hasChannel(channelId)) {
-    this._worker.createChannel(channelId);
+  if (!this._channelList.hasChannel(channelId)) {
+    this._channelList.createChannel(channelId);
   }
-  var channel = this._worker.getChannel(channelId);
+  var channel = this._channelList.getChannel(channelId);
   clearTimeout(channel.closeTimeout);
   channel.addSubscriber(new Subscriber(this, data));
   _.each(channel.getMessages(), function(message) {
@@ -86,15 +86,15 @@ WorkerConnection.prototype.subscribe = function(channelId, data, msgStartTime) {
  */
 WorkerConnection.prototype.unsubscribe = function(channelId) {
   this.removeChannelId(channelId);
-  if (!this._worker.hasChannel(channelId)) {
+  if (!this._channelList.hasChannel(channelId)) {
     return;
   }
-  var channel = this._worker.getChannel(channelId);
+  var channel = this._channelList.getChannel(channelId);
   WorkerUtil.sendUpUnsubscribe(channelId, this.id);
 
   channel.unsubscribe(this);
   if (!channel.hasSubscribers()) {
-    this._worker.delayedCloseChannel(channelId);
+    this._channelList.delayedCloseChannel(channelId);
   }
 };
