+++ lib/worker.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var redis = require('redis');
 var sockjs = require('sockjs');
 var validator = require('validator');
-var Channel = require('./channel');
+var ChannelList = require('./channel-list');
 var WorkerUtil = require('./worker-util');
 var WorkerConnection = require('./worker-connection');
 
@@ -13,8 +13,8 @@ var WorkerConnection = require('./worker-connection');
  * @constructor
  */
 function Worker(port, sockjsClientUrl, sslOptions) {
-  /** @type {Object} */
-  this._channels = {};
+  /** @type {ChannelList} */
+  this._channelList = new ChannelList();
   /** @type {Array} */
   this._connections = [];
 
@@ -27,31 +27,6 @@ function Worker(port, sockjsClientUrl, sslOptions) {
 
 Worker.prototype.stop = function() {
   this._httpServer.close();
-};
-
-/**
- * @param {String} channelId
- * @returns {Boolean}
- */
-Worker.prototype.hasChannel = function(channelId) {
-  return !!this._channels[channelId];
-};
-
-/**
- * @param {String} channelId
- * @returns {Channel}
- */
-Worker.prototype.getChannel = function(channelId) {
-  return this._channels[channelId];
-};
-
-/**
- * @param {String} channelId
- * @returns {Channel} channel
- */
-Worker.prototype.createChannel = function(channelId) {
-  this._channels[channelId] = new Channel(channelId);
-  return this._channels[channelId];
 };
 
 /**
@@ -83,7 +58,7 @@ Worker.prototype._listen = function(port) {
       console.error('Empty WebSocket connection');
       return;
     }
-    self._addConnection(new WorkerConnection(connection, self));
+    self._addConnection(new WorkerConnection(connection, self._channelList));
   });
 
   this._sockjsServer.installHandlers(this._httpServer);
@@ -113,10 +88,10 @@ Worker.prototype._removeConnection = function(connection) {
  * @param {Object} data
  */
 Worker.prototype.sendDownPublish = function(channelId, event, data) {
-  var channel = this.getChannel(channelId);
+  var channel = this._channelList.getChannel(channelId);
   if (!channel) {
-    channel = this.createChannel(channelId);
-    this.delayedCloseChannel(channelId);
+    channel = this._channelList.createChannel(channelId);
+    this._channelList.delayedCloseChannel(channelId);
   }
   var content = {channel: channelId, event: event, data: data};
   channel.addMessage(content);
@@ -130,27 +105,12 @@ Worker.prototype.sendDownPublish = function(channelId, event, data) {
  */
 Worker.prototype._getChannelsData = function() {
   var channelsData = {};
-  _.each(this._channels, function(channel, channelId) {
+  _.each(this._channelList.getChannels(), function(channel, channelId) {
     channelsData[channelId] = _.map(channel.getSubscribers(), function(subscriber) {
       return {clientKey: subscriber.connection.id, data: subscriber.data, subscribeStamp: subscriber.createStamp};
     });
   });
   return channelsData;
-};
-
-/**
- * @param {String} channelId
- */
-Worker.prototype.delayedCloseChannel = function(channelId) {
-  var channel = this.getChannel(channelId);
-  if (channel) {
-    if (channel.closeTimeout) {
-      clearTimeout(channel.closeTimeout);
-    }
-    channel.closeTimeout = setTimeout(function() {
-      delete this._channels[channelId];
-    }.bind(this), 10000);
-  }
 };
 
 /**
